+++ LibCarla/source/carla/rpc/Command.h
@@ -132,6 +132,16 @@ namespace rpc {
       MSGPACK_DEFINE_ARRAY(actor, impulse);
     };
 
+    struct ApplyForce : CommandBase<ApplyForce> {
+      ApplyForce() = default;
+      ApplyForce(ActorId id, const geom::Vector3D &value)
+        : actor(id),
+          force(value) {}
+      ActorId actor;
+      geom::Vector3D force;
+      MSGPACK_DEFINE_ARRAY(actor, force);
+    };
+
     struct ApplyAngularImpulse : CommandBase<ApplyAngularImpulse> {
       ApplyAngularImpulse() = default;
       ApplyAngularImpulse(ActorId id, const geom::Vector3D &value)
@@ -140,6 +150,16 @@ namespace rpc {
       ActorId actor;
       geom::Vector3D impulse;
       MSGPACK_DEFINE_ARRAY(actor, impulse);
+    };
+
+    struct ApplyTorque : CommandBase<ApplyTorque> {
+      ApplyTorque() = default;
+      ApplyTorque(ActorId id, const geom::Vector3D &value)
+        : actor(id),
+          torque(value) {}
+      ActorId actor;
+      geom::Vector3D torque;
+      MSGPACK_DEFINE_ARRAY(actor, torque);
     };
 
     struct SetSimulatePhysics : CommandBase<SetSimulatePhysics> {
@@ -189,7 +209,9 @@ namespace rpc {
         ApplyTargetVelocity,
         ApplyTargetAngularVelocity,
         ApplyImpulse,
+        ApplyForce,
         ApplyAngularImpulse,
+        ApplyTorque,
         SetSimulatePhysics,
         SetAutopilot,
         SetVehicleLightState>;